Refactored project
Using https://github.com/crsandeep/simple-react-full-stack
Fixes #34
