+++ .gitignore
@@ -1,5 +1,5 @@
 # dependencies
-/node_modules
+node_modules
 .env
 
 # testing
@@ -7,6 +7,7 @@
 
 # production
 /build
+dist
 
 # misc
 .DS_Store
