--- client/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "fmb-app",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "jss": "^10.0.0-alpha.12",
-    "radium": "^0.25.1",
-    "react": "^16.8.4",
-    "react-dom": "^16.8.4",
-    "react-scripts": "2.1.8",
-    "typescript": "^3.3.3333"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
-}
