+++ package.json
@@ -1,14 +1,62 @@
 {
   "name": "findmybork",
-  "main": "server/app/server.js",
+  "version": "2.0.0",
+  "description": "Find My Bork",
+  "main": "src/server/server.js",
   "scripts": {
-    "start": "npm start --prefix ./server",
-    "heroku-prebuild": "cd server && npm install && npm run build",
-    "heroku-postbuild": "cd client && npm install && npm run build"
+    "build": "webpack --mode production",
+    "start": "npm run build && node --experimental-modules src/server/server.mjs ",
+    "client": "webpack-dev-server --mode development --devtool inline-source-map --hot",
+    "server": "nodemon src/server/server.js",
+    "dev": "concurrently \"npm run server\" \"npm run client\""
   },
   "engines": {
     "node": "10.3.0",
     "npm": "6.x"
   },
-  "license": "MIT"
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/fmb-app/fmb.git"
+  },
+  "author": "Group 50",
+  "license": "MIT",
+  "dependencies": {
+    "babel-polyfill": "^6.26.0",
+    "express": "^4.16.4",
+    "react": "^16.8.4",
+    "react-dom": "^16.8.4",
+    "systemet": "^1.0.0",
+    "jss": "^10.0.0-alpha.12",
+    "radium": "^0.25.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "babel-eslint": "^10.0.0",
+    "babel-loader": "^8.0.0",
+    "clean-webpack-plugin": "^1.0.0",
+    "concurrently": "^4.0.0",
+    "css-loader": "^2.0.0",
+    "dotenv": "^6.2.0",
+    "eslint": "^5.0.0",
+    "eslint-config-airbnb": "^17.0.0",
+    "eslint-plugin-import": "^2.11.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.7.0",
+    "file-loader": "^3.0.0",
+    "html-webpack-plugin": "^3.2.0",
+    "nodemon": "^1.17.3",
+    "style-loader": "^0.23.0",
+    "volleyball": "^1.5.1",
+    "url-loader": "^1.0.1",
+    "webpack": "^4.5.0",
+    "webpack-cli": "^3.0.8",
+    "webpack-dev-server": "^3.1.3"
+  },
+  "bugs": {
+    "url": "https://github.com/fmb-app/fmb/issues"
+  },
+  "homepage": "https://github.com/fmb-app/fmb#readme"
 }
