+++ src/client/components/BottomNavBar/BottomNavBar.js
