+++ src/client/components/MidSection/MidSection.js
